backend/requirements.txt
# RSE Backend Dependencies — pinned for reproducibility
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9

# Vercel ASGI adapter — wraps FastAPI for Vercel/Lambda serverless deployment
mangum==0.17.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
